refactor(providers): move FTP/HTTPS transport into providers/transport.py
Pulled download_ftp_urls, download_http_urls, and their helpers
(_open_ftp_connection, _walk_ftp_tree, _list_ftp_repo_files,
_download_one_ftp_path, _download_ftp_paths_serial/_parallel,
_http_download_one, _parallel_download, _local_path_for_url) out of the
Files class verbatim into providers/transport.py.

Files keeps a shim staticmethod for each function that does a lazy
import and delegates, so existing test patches like
patch.object(Files, 'download_ftp_urls') keep intercepting calls.

No behaviour change. Test suite green.
